Performance improvements
Improved line contribution calculations and in each step of the set
cover algorithm we check if the best attribute covers more lines than
what remains to be covered. In that case we calculate the sum of the
remaining lines.
If the attribute covers only a few lines we remove the contribution of
the covered lines.
The objetive is to reduce the number of lines read from the dataset.